+++ packages/react-magma-dom/src/components/Badge/Badge.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { Badge } from '.';
+
+export default {
+  component: Badge,
+  title: 'Badge',
+};
+
+export const Default = () => {
+  return <Badge>Badge</Badge>;
+};