+++ packages/react-magma-dom/src/components/ProgressBar/ProgressBar.stories.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+import { ProgressBar } from '.';
+
+export default {
+  component: ProgressBar,
+  title: 'ProgressBar',
+};
+
+export const Default = () => {
+  return <ProgressBar percentage={25} />;
+};